requirements.txt
SQLAlchemy==2.0.25
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
supabase==2.3.4

